api/serializers/category.py
from rest_framework import serializers

import main.models as models


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Category
        fields = "__all__"