requirements.txt
django
django-bootstrap-v5
pillow
drf-yasg
djangorestframework
djangorestframework-simplejwt